examples/requirements_djangocms.txt
django-cms==3.0.6
#django-cms==2.4.1
djangocms-picture
djangocms-snippet
djangocms-text-ckeditor
django-mptt
djangocms-admin-style